+++ src/main/java/com/starkbank/Settings.java
@@ -1,0 +1,8 @@
+package com.starkbank;
+
+
+public abstract class Settings{
+    public static Project user = null;
+    public static String userAgentOverride = null;
+    public static String language = "en-US";
+}