+++ src/main/java/com/starkbank/User.java
@@ -6,9 +6,6 @@ import com.starkbank.utils.Resource;
 
 
 public abstract class User extends Resource{
-    public static Project defaultUser = null;
-    public static String userAgentOverride = null;
-
     public final String pem;
     public final String environment;