+++ src/test/java/TestBalance.java
@@ -1,5 +1,5 @@
 import com.starkbank.Balance;
-import com.starkbank.User;
+import com.starkbank.Settings;
 import org.junit.Test;
 import org.junit.Assert;
 
@@ -8,7 +8,7 @@ public class TestBalance {
 
     @Test
     public void testGet() throws Exception {
-        User.defaultUser = utils.User.defaultProject();
+        Settings.user = utils.User.defaultProject();
         Balance balance = Balance.get();
         System.out.println(balance);
         Assert.assertNotNull(balance.id);